pyproject.toml
[tool.poetry]
name = "seeonline"
version = "0.1.0"
description = ""
authors = ["Fe4ester <xdxdxd4305@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.6"
psycopg2-binary = "^2.9.10"
telethon = "^1.39.0"
celery = {extras = ["redis"], version = "^5.4.0"}
django-celery-beat = "^2.7.0"
djangorestframework = "^3.15.2"
django-filter = "^25.1"
drf-spectacular = "^0.28.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
